+++ Sources/VCore/Extensions/SwiftUI/View.OnSizeChange.swift
@@ -1,5 +1,5 @@
 //
-//  View.ReadSize.swift
+//  View.OnSizeChange.swift
 //  VCore
 //
 //  Created by Vakhtang Kontridze on 10/28/21.
@@ -7,7 +7,7 @@
 
 import SwiftUI
 
-// MARK: - View Read Size
+// MARK: - View On Size Change
 extension View {
     /// Reads `View` size and calls an on-change block.
     ///
@@ -16,12 +16,12 @@ extension View {
     ///     var body: some View {
     ///         VStack(content: {
     ///             Color.accentColor
-    ///                 .readSize(onChange: { size = $0 })
+    ///                 .onSizeChange(perform: { size = $0 })
     ///         })
     ///     }
     ///
-    public func readSize(
-        onChange completion: @escaping (CGSize) -> Void
+    public func onSizeChange(
+        perform action: @escaping (CGSize) -> Void
     ) -> some View {
         background(
             GeometryReader(content: { proxy in
@@ -29,7 +29,7 @@ extension View {
                     .preference(key: SizePreferenceKey.self, value: proxy.frame(in: .local).size)
             })
         )
-            .onPreferenceChange(SizePreferenceKey.self, perform: completion)
+            .onPreferenceChange(SizePreferenceKey.self, perform: action)
     }
 }